+++ app/src/main/java/com/shub39/grit/viewmodels/HabitViewModel.kt
@@ -101,7 +101,7 @@ class HabitViewModel(
         habitStatusJob?.cancel()
         habitStatusJob = viewModelScope.launch {
             combine(
-                repo.getHabitStatus(),
+                repo.getHabitsWithAnalytics(),
                 repo.getCompletedHabitIds()
             ) { habits, completedHabits ->
                 _state.update {